+++ package.json
@@ -27,5 +27,6 @@
   "engines": {
     "node": "16.x",
     "npm": "8.x"
-  }
+  },
+  "type": "module"
 }
